First cut at a landing page
This is a VERY ROUGH take on the landing page. The basic idea is we want
to present 3 licenses on the home page and just enough info for people
to choose one of the three. But we'll provide links for folks who need
to dig deeper and choose a license that better fits their needs.

The design right now is blatantly stolen from KOD. But that's just to
make the early draft look pretty. That's not the intended final design.