licenses: pre-format bsd license texts
The license text is shown as preformatted text; therefore we
need to wrap it to a sane length (like 80 characters) to
keep it readable.

In addition, we must do any formatting or indentation in the
source file. Both BSD licenses contain a list of conditions
that should be indented like a list, with blank lines in
between.

We could also potentially number the clauses; FreeBSD does
in its copy, but the OSI versions do not.