AGPL: Highlight the Affero Feature
This is a follow up to #197

A new point in the "required" section of the AGPL
highlights the difference to the plain GPL.
It states that running a network service of the software
triggers the disclose-source requirement, e.g. in SaaS use cases.

This is my personal interpretation of the paragraph 13 AGPL, IANAL.

Comments are welcome!