Remove false assertion from goals
Whether in civil rights, human rights, land rights, "intellectual property" rights (as in this case) or any other rights, the act of deciding who ought to be granted which rights in which circumstances is not only inherently political, it is the very essence of politics. Therefore, it is preposterous for choosealicense.com, if it is to be meaningful, to aim to be "Politics Free".