Remove flash detection
Accurate flash detection is a little more complicated than this. The
current code causes an error on IE 10.

This removes the attempt to detect if flash is available. Instead, it
attempts to initialize ZeroClipboard and sets up the fallback.
ZeroClipboard will be a no-op if flash is not available, in which case
the fallback will already be configured. If flash is available, it will
put a flash object over the button and the fallback will never get
called.