always usually evaluates to false
Remove "If you have questions or issues, it is always best to consult
a legal professional."

It's not always best, especially considering it's costly and for
most finding one knowledgable about open source licensing will be
difficult.

If it were always best, the site should consist of

<h1>Consult a legal professional.</h1>

But seriously, claiming one should consult a legal professional is
no way to make choosing an open source license non-scary.