Restore TLDRLegal as "additional resource"
I noticed that TLDRLegal was taken down due to concerns around spam -- we did a big sweep of content in between then and now.

Would be cool to be back on this page, esp given the history of us and ChooseALicense! Feel free to reach out at kevin (at) tldrlegal.com beforehand as well.