expansion/clarification of non-license permissions
I made some suggestions to this documentation today because I have been wondering about these ambiguities for years and am recently, trying to convince a repo maintainer to add a license to his repo which has over 1000 regular users; but this document (even the section: "Ask the maintainers nicely to add a license") offers nothing convincing to that end.  In fact, it is barely self-consistent, and the GitHub TOS is so terse on the topic that it is not at all clear what this document implies specifically for GitHub users.

The one paragraph states that GitHub public repos are forkable (and actually download-able) regardless of the license or lack thereof.  Then the very next paragraph states that without a license users may not use the material in ANY way.  This avoids blatant contradiction only by omitting that, strictly speaking, without a license the rights do not exist to copy or fork either.

I hope this edit will serve to inform users until perhaps GitHub more clearly defines the limits of the permissions granted in section F1 of the TOS.  Until then, I hope that the drafters of the GitHub TOS would read this PR mindfully and note that it raises some important issues.