check-approval: Fix missing license_ids variable
Avoid:

   $ ./script/check-approval MIT
   ./script/check-approval:55:in `<main>': undefined local variable or method `license_ids' for main:Object (NameError)

The license_ids line was added in e505eb8f (check if license is aleady
a license, 2016-01-18, #318), but license_ids was removed from the
helper in b99e7ab0 (replace 'id' variables with 'spdx_lcase' to
minimize confusion, 2016-06-01, #424).  This commit restores the old
code locally, since this script is the only consumer.