Remove "use" list of things that require a license
At least in the United States, you can use a copyrighted work without having a license. 17 USC 106 lists the things you need a license to do and it does not include the ordinary use of a lawfully-acquired copy of a work. If this wasn't true, you couldn't even read a book you owned unless that book included an explicit license (and most don't).