Context; clarifying poor wording; noob friendly
Provided context to the beginning of the list for noobs in licensing. Clarified poor word choices with definitions and more common terms; reversed the list to be more intuitive for noobs to licensing, it begins with licenses with fewer stipulations. Make this an open door for people who have never read about licensing.