Improve Language Comprehension
These changes help non-English speakers to read along.

Changing "but" to "and" suggests a correct continual pathos whilst adding clarity.
"But" is used to suggest a hard shift in mindset. It was misused.

Changing "them" to "those" clarified that the sentence is about fonts and not about people setting the license.
"Them" is used as shorthand for humans. It will confuse the reader.