Spanish version of the website
Hi, I am Francisco and I am from Argentina, I did not find such a complete repository of licenses in spanish so I decided to do something for the community and make a translated version of yours. 
This version is currently being hosted by me in eligeunalicencia.com . 
But I will take it down soon, I would like to merge it with our page and make a section for the spanish speakers if you do agree.